Select a random image from the set of images in each grouping
Fix the image selection to not always pick the first image from the
sorted list. For each group (tags and categories), find all of the
members that have an available image, shuffle that collection, and pick
the first element of the randomized collection. This should solve the
featured image always being the same tag.