Add new action to run before Facebook Auto Publish
The Facebook Auto Publish plugin, if it finds a feature image, will use
that when posting to Facebook. The existing code ran too late most of
the time. Use the transition_post_status action with a lower priority to
run first to set a featured image when a post is published.